Inline (but don't yet simplify) `Server.Call`
Simplification will be easier once the rest of the server infrastructure is
also inlined.
