+++ test-grapesy/Main.hs
@@ -40,18 +40,22 @@ import Network.HTTP2.Server qualified as HTTP2
 import Network.Run.TCP (runTCPServer)
 import Network.Run.TCP qualified as Run
 import Network.Socket
+import Network.HTTP.Types qualified as HTTP
+import Network.HTTP2.Internal qualified as HTTP2
 
 import Network.GRPC.Common
 import Network.GRPC.Common.StreamElem qualified as StreamElem
-import Network.GRPC.Server.Call qualified as Server
 import Network.GRPC.Server.Connection (Connection(..), withConnection)
 import Network.GRPC.Server.Connection qualified as Connection
 import Network.GRPC.Server.Context (ServerParams)
 import Network.GRPC.Server.Context qualified as Context
+import Network.GRPC.Server.Session
 import Network.GRPC.Spec
 import Network.GRPC.Util.Concurrency
+import Network.GRPC.Util.HTTP2.Stream (ClientDisconnected(..))
 import Network.GRPC.Util.Parser
 import Network.GRPC.Util.Session qualified as Session
+import Network.GRPC.Util.Session.Server qualified as Server
 
 -- ========================================================================== --
 
@@ -239,10 +243,10 @@ runServer (ServerConfig port) server =
 
 data RpcHandler = forall rpc. IsRPC rpc => RpcHandler {
       -- | Handler proper
-      runRpcHandler :: Server.Call rpc -> IO ()
+      runRpcHandler :: ServerCall rpc -> IO ()
     }
 
-mkRpcHandler :: IsRPC rpc => Proxy rpc -> (Server.Call rpc -> IO ()) -> RpcHandler
+mkRpcHandler :: IsRPC rpc => Proxy rpc -> (ServerCall rpc -> IO ()) -> RpcHandler
 mkRpcHandler _ = RpcHandler
 
 type HandlerMap = [(Path, RpcHandler)]
@@ -253,7 +257,7 @@ constructHandlerMap = map (\h -> (handlerPath h, h))
 handlerPath :: RpcHandler -> Path
 handlerPath RpcHandler{runRpcHandler} = aux runRpcHandler
   where
-    aux :: forall rpc. IsRPC rpc => (Server.Call rpc -> IO ()) -> Path
+    aux :: forall rpc. IsRPC rpc => (ServerCall rpc -> IO ()) -> Path
     aux _ = rpcPath (Proxy @rpc)
 
 -- ========================================================================== --
@@ -285,7 +289,7 @@ handleRequest params handlers conn = do
                     Just h  -> h
                     Nothing -> error "withHandler: unknown path"
     case handler of
-      RpcHandler h -> Server.acceptCall params conn h
+      RpcHandler h -> serverAcceptCall params conn h
   where
     path :: Path
     path = Connection.path conn
@@ -295,6 +299,465 @@ handleRequest params handlers conn = do
           Context.serverDebugTracer
         $ Context.params
         $ connectionContext conn
+
+-- ========================================================================== --
+
+{-------------------------------------------------------------------------------
+  Definition
+-------------------------------------------------------------------------------}
+
+data ServerCall rpc = IsRPC rpc => ServerCall {
+      -- | Server state (across all calls)
+      callSession :: ServerSession rpc
+
+      -- | Bidirectional channel to the client
+    , callChannel :: Session.Channel (ServerSession rpc)
+
+      -- | Request metadata
+      --
+      -- This is filled once we get the request headers
+    , callRequestMetadata :: TMVar [CustomMetadata]
+
+      -- | Response metadata
+      --
+      -- Metadata to be included when we send the initial response headers.
+      -- Can be updated until the first message (see 'callFirstMessage').
+      --
+      -- Defaults to the empty list.
+    , callResponseMetadata :: TVar [CustomMetadata]
+
+      -- | What kicked off the response?
+      --
+      -- This is empty until the response /has/ in fact been kicked off.
+    , callResponseKickoff :: TMVar Kickoff
+    }
+
+-- | What kicked off the response?
+--
+-- When the server handler starts, we do not immediately initate the response
+-- to the client, because the server might not have decided the initial response
+-- metadata yet (indeed, it might need wait for some incoming messages from the
+-- client before it can compute that metadata). We therefore wait until we
+-- send the initial response headers, until the handler..
+--
+-- 1. explicitly calls 'initiateResponse'
+-- 2. sends the first message using 'sendOutput'
+-- 3. initiates and immediately terminates the response using 'sendTrailersOnly'
+--
+-- We only need distinguish between (1 or 2) versus (3), corresponding precisely
+-- to the two constructors of 'FlowStart'.
+data Kickoff =
+    KickoffRegular
+  | KickoffTrailersOnly TrailersOnly
+  deriving (Show)
+
+{-------------------------------------------------------------------------------
+  Open a call
+-------------------------------------------------------------------------------}
+
+-- | Accept incoming call
+--
+-- If an exception is thrown during call setup, we will send an error response
+-- to the client (and then rethrow the exception).
+serverAcceptCall :: forall rpc.
+     (IsRPC rpc, HasCallStack)
+  => Context.ServerParams
+  -> Connection
+  -> (ServerCall rpc -> IO ())
+  -> IO ()
+serverAcceptCall params conn k = do
+    callRequestMetadata  <- newEmptyTMVarIO
+    callResponseMetadata <- newTVarIO []
+    callResponseKickoff  <- newEmptyTMVarIO
+
+    let setupResponseChannel :: IO (Session.Channel (ServerSession rpc))
+        setupResponseChannel =
+            Session.setupResponseChannel
+              callSession
+              (contramap (Context.ServerDebugMsg @rpc) tracer)
+              (Connection.connectionClient conn)
+              (   handle sendErrorResponse
+                . mkOutboundHeaders
+                    callRequestMetadata
+                    callResponseMetadata
+                    callResponseKickoff
+              )
+
+        handlerTeardown ::
+             ServerCall rpc
+          -> Either SomeException ()
+          -> IO ()
+        handlerTeardown call@ServerCall{callChannel} mRes = do
+            mUnclean <-
+              case mRes of
+                Right () ->
+                  -- Handler terminated successfully
+                  -- (but it might not have sent the final message, see below)
+                  Session.close callChannel $ ExitCaseSuccess ()
+                Left err -> do
+                  -- The handler threw an exception. /Try/ to tell the client
+                  -- (but see discussion of 'forwardException')
+                  traceWith (Context.serverExceptionTracer params) err
+                  forwardException call err
+                  Session.close callChannel $ ExitCaseException err
+
+            case mUnclean of
+              Nothing  -> return () -- Channel was closed cleanly
+              Just err ->
+                -- An unclean shutdown can have 3 causes:
+                --
+                -- 1. We failed to set up the call.
+                --
+                --    If the failure is due to network comms, we definitely have
+                --    no way of communicating with the client. If the failure
+                --    was due to a setup failure in 'mkOutboundHeaders', then
+                --    the client has /already/ been notified of the problem.
+                --
+                -- 2. We lost communication during the call.
+                --
+                --    In this case we also have no way of telling the client
+                --    that something went wrong.
+                --
+                -- 3. The handler failed to properly terminate the communication
+                --    (send the final message and call 'waitForOutbound').
+                --
+                --    This is the trickiest case. We don't really know what
+                --    state the handler left the channel in; for example, we
+                --    might have killed the thread halfway through sending a
+                --    message.
+                --
+                -- So the only thing we really can do here is log the error,
+                -- and nothing else. (We should not rethrow it, as doing so will
+                -- cause http2 to reset the stream, which is not always the
+                -- right thing to do (for example, in case (1)).
+                traceWith (Context.serverExceptionTracer params) $
+                  toException err
+
+    let handler ::
+             Session.Channel (ServerSession rpc)
+          -> (forall a. IO a -> IO a)
+          -> IO ()
+        handler callChannel unmask = do
+            mRes <- try $ unmask $ k call
+            handlerTeardown call mRes
+          where
+            call :: ServerCall rpc
+            call = ServerCall{
+                callSession
+              , callChannel
+              , callRequestMetadata
+              , callResponseMetadata
+              , callResponseKickoff
+              }
+
+    runHandler setupResponseChannel handler
+  where
+    callSession :: ServerSession rpc
+    callSession = ServerSession
+
+    tracer :: Tracer IO Context.ServerDebugMsg
+    tracer = Context.serverDebugTracer $ Context.params $ connectionContext conn
+
+    mkOutboundHeaders ::
+         TMVar [CustomMetadata]
+      -> TVar [CustomMetadata]
+      -> TMVar Kickoff
+      ->     Session.FlowStart (ServerInbound  rpc)
+      -> IO (Session.FlowStart (ServerOutbound rpc))
+    mkOutboundHeaders requestMetadataVar
+                      responseMetadataVar
+                      responseKickoffVar
+                      requestStart
+                    = do
+        traceIO "mkOutboundHeaders 1"
+        -- Make request metadata available (see 'getRequestMetadata')
+        atomically $ putTMVar requestMetadataVar $
+                       requestMetadata inboundHeaders
+        traceIO "mkOutboundHeaders 2"
+
+        -- Wait for kickoff (see 'Kickoff' for discussion)
+        kickoff <- atomically $ readTMVar responseKickoffVar
+        traceIO "mkOutboundHeaders 3"
+
+        -- Get response metadata (see 'setResponseMetadata')
+        responseMetadata <- atomically $ readTVar responseMetadataVar
+        traceIO "mkOutboundHeaders 4"
+
+        -- Session start
+        case kickoff of
+          KickoffRegular -> do
+            traceIO $ "mkOutboundHeaders 5"
+            return $ Session.FlowStartRegular $ OutboundHeaders {
+                outHeaders = ResponseHeaders {
+                    responseMetadata = responseMetadata
+                  }
+              }
+          KickoffTrailersOnly trailers -> do
+            traceIO $ "mkOutboundHeaders 6"
+            return $ Session.FlowStartNoMessages trailers
+      where
+        inboundHeaders :: RequestHeaders
+        inboundHeaders =
+            case requestStart of
+              Session.FlowStartRegular    headers  -> inbHeaders headers
+              Session.FlowStartNoMessages trailers ->            trailers
+
+    -- | Send response when 'mkOutboundHeaders' fails
+    --
+    -- TODO: Some duplication with 'forwardException' (and 'withHandler').
+    sendErrorResponse :: forall x. SomeException -> IO x
+    sendErrorResponse err = do
+        traceWith tracer $ Context.ServerDebugAcceptFailed err
+        Server.respond (connectionClient conn) $
+          HTTP2.responseNoBody
+            HTTP.ok200 -- gRPC uses HTTP 200 even when there are gRPC errors
+            (buildTrailersOnly $ TrailersOnly $ ProperTrailers {
+                trailerGrpcStatus  = GrpcError GrpcUnknown
+                -- TODO: Potential security concern here
+                -- (showing the exception)?
+              , trailerGrpcMessage = Just $ Text.pack (show err)
+              , trailerMetadata    = []
+              })
+        throwIO err
+
+-- | Run the handler
+--
+-- http2 will kill the handler when the client disappears, but we don't want
+-- that behaviour: we want the handler to be able to terminate cleanly. We
+-- therefore run the handler in a separate thread, and wait for that thread to
+-- terminate. If we are interrupted while we wait, it depends on the
+-- interruption:
+--
+-- * If we are interrupted by 'HTTP2.KilledByHttp2ThreadManager', it means
+--   we got disconnected from the client. In this case, we shut down the channel
+--   (if it's not already shut down); /if/ the handler at this tries to
+--   communicate with the client, an exception will be raised. However, the
+--   handler /can/ terminate cleanly and, of course, typically will.
+--   Importantly, this avoids race conditions: even if the server and the client
+--   agree that no further communication takes place (the client sent their
+--   final message, the server sent the trailers), without the indireciton of
+--   this additional thread it can still happen that http2 kills the handler
+--   (after the client disconnects) before the handler has a chance to
+--   terminate.
+-- * If we are interrupted by another kind of asynchronous exception, we /do/
+--   kill the handler (this might for example be a timeout).
+--
+-- This is in line with the overall design philosophy of communication in this
+-- library: exceptions will only be raised synchronously when communication is
+-- attempted, not asynchronously when we notice a problem.
+runHandler ::
+     IO (Session.Channel (ServerSession rpc))
+  -> (Session.Channel (ServerSession rpc) -> (forall a. IO a -> IO a) -> IO ())
+  -> IO ()
+runHandler setupResponseChannel handler = do
+    mask $ \unmask -> do
+      -- This sets up the channel but no response is sent yet
+      callChannel   <- setupResponseChannel
+      handlerThread <- asyncWithUnmask $ handler callChannel
+
+      let loop :: IO ()
+          loop = do
+             -- We do not need to use 'waitCatch': the handler installs its
+             -- own exception handler and will not throw any exceptions.
+             mRes <- try $ unmask $ wait handlerThread
+             case mRes of
+               Right ()  -> return ()
+               Left  err -> do
+                 case fromException err of
+                   Just (HTTP2.KilledByHttp2ThreadManager mErr) -> do
+                     let exitReason :: ExitCase ()
+                         exitReason =
+                           case mErr of
+                             Nothing -> ExitCaseSuccess ()
+                             Just exitWithException ->
+                               ExitCaseException . toException $
+                                 ClientDisconnected exitWithException
+                     _mAlreadyClosed <- Session.close callChannel exitReason
+                     loop
+                   Nothing -> do
+                     cancelWith handlerThread err
+                     throwIO err
+
+      loop
+
+-- | Process exception thrown by a handler
+--
+-- Trace the exception and forward it to the client.
+--
+-- The attempt to forward it to the client is a best-effort only:
+--
+-- * The nature of the exception might mean that we we cannot send anything to
+--   the client at all.
+-- * It is possible the exception was thrown /after/ the handler already send
+--   the trailers to the client.
+--
+-- We therefore catch and suppress all exceptions here.
+forwardException :: ServerCall rpc -> SomeException -> IO ()
+forwardException call err = do
+    handle ignoreExceptions $
+      sendProperTrailers call trailers
+  where
+    trailers :: ProperTrailers
+    trailers
+      | Just (err' :: GrpcException) <- fromException err
+      = grpcExceptionToTrailers err'
+
+      -- TODO: There might be a security concern here (server-side exceptions
+      -- could potentially leak some sensitive data).
+      | otherwise
+      = ProperTrailers {
+            trailerGrpcStatus  = GrpcError GrpcUnknown
+          , trailerGrpcMessage = Just $ Text.pack (show err)
+          , trailerMetadata    = []
+          }
+
+    -- See discussion above.
+    ignoreExceptions :: SomeException -> IO ()
+    ignoreExceptions _ = return ()
+
+{-------------------------------------------------------------------------------
+  Open (ongoing) call
+-------------------------------------------------------------------------------}
+
+-- | Receive RPC input from the client
+--
+-- We do not return trailers, since gRPC does not support sending trailers from
+-- the client to the server (only from the server to the client).
+serverRecvInput :: forall rpc. HasCallStack => ServerCall rpc -> IO (StreamElem NoMetadata (Input rpc))
+serverRecvInput call =
+    catch go $ \e@BlockedIndefinitelyOnSTM{} -> do
+      traceIO $ "Uhoh.. got the dreaded " ++ show e
+      throwIO e
+      --threadDelay 1_000_000
+      --traceIO $ "Trying again!"
+      --go
+  where
+    go :: IO (StreamElem NoMetadata (Input rpc))
+    go = atomically $ recvInputSTM call
+
+-- | Send RPC output to the client
+--
+-- This will send a @grpc-status@ of @0@ to the client; for anything else (i.e.,
+-- to indicate something went wrong), the server handler should throw a
+-- 'GrpcException' (the @grapesy@ client API treats this the same way: a
+-- @grpc-status@ other than @0@ will be raised as a 'GrpcException').
+--
+-- This is a blocking call if this is the final message (i.e., the call will
+-- not return until the message has been written to the HTTP2 stream).
+serverSendOutput :: ServerCall rpc -> StreamElem [CustomMetadata] (Output rpc) -> IO ()
+serverSendOutput call msg = do
+    _updated <- initiateResponse call
+    atomically $ sendOutputSTM call msg
+    StreamElem.whenDefinitelyFinal msg $ \_ -> waitForOutbound call
+
+{-------------------------------------------------------------------------------
+  Low-level API
+
+  'initiateResponse' and 'sendOutputSTM' /MUST NOT/ live in the same
+  transaction: in order to send a message, we need to know if the call is in
+  "regular" flow state or "trailers only" flow state, which we only know after
+  'initiateResponse'; so if these live in the same transaction, we will
+  deadlock.
+-------------------------------------------------------------------------------}
+
+-- | STM version of 'recvInput'
+--
+-- This is a low-level API; most users can use 'recvInput' instead.
+--
+-- Most server handlers will deal with single clients, but in principle
+-- 'recvInputSTM' could be used to wait for the first message from any number of
+-- clients.
+recvInputSTM :: forall rpc.
+     HasCallStack
+  => ServerCall rpc
+  -> STM (StreamElem NoMetadata (Input rpc))
+recvInputSTM ServerCall{callChannel} =
+    first ignoreTrailersOnly <$> Session.recv callChannel
+  where
+    ignoreTrailersOnly ::
+         Either RequestHeaders NoMetadata
+      -> NoMetadata
+    ignoreTrailersOnly _ = NoMetadata
+
+-- | STM version of 'sendOutput'
+--
+-- This is a low-level API; most users can use 'sendOutput' instead.
+--
+-- If you choose to use 'sendOutputSTM' instead, you have two responsibilities:
+--
+-- * You must call 'initiateResponse' before calling 'sendOutputSTM';
+--   'sendOutputSTM' will throw 'ResponseNotInitiated' otherwise.
+-- * You must call 'waitForOutbound' after sending the final message (and before
+--   exiting the scope of 'acceptCall'); if you don't, you risk that the HTTP2
+--   stream is cancelled.
+--
+-- Implementation note: we cannot call 'waitForOutbound' /in/ 'acceptCall': if
+-- the handler for whatever reason never writes the final message, then such a
+-- call would block indefinitely.
+sendOutputSTM :: ServerCall rpc -> StreamElem [CustomMetadata] (Output rpc) -> STM ()
+sendOutputSTM ServerCall{callChannel, callResponseKickoff} msg = do
+    mKickoff <- tryReadTMVar callResponseKickoff
+    case mKickoff of
+      Just _  -> Session.send callChannel (first mkTrailers msg)
+      Nothing -> throwSTM $ ResponseNotInitiated
+  where
+    mkTrailers :: [CustomMetadata] -> ProperTrailers
+    mkTrailers metadata = ProperTrailers {
+          trailerGrpcStatus  = GrpcOk
+        , trailerGrpcMessage = Nothing
+        , trailerMetadata    = metadata
+        }
+
+-- | Initiate the response
+--
+-- See 'sendOutputSTM' for discusison.
+--
+-- Returns 'False' if the response was already initiated.
+initiateResponse :: ServerCall rpc -> IO Bool
+initiateResponse ServerCall{callResponseKickoff} =
+    atomically $ tryPutTMVar callResponseKickoff KickoffRegular
+
+-- | Wait for all outbound messages to have been processed
+--
+-- This function /must/ be called before leaving the scope of 'acceptCall'.
+-- However, 'sendOutput' will call 'waitForOutbound' when the last message is
+-- sent, so you only need to worry about this if you are using 'sendOutputSTM'.
+waitForOutbound :: ServerCall rpc -> IO ()
+waitForOutbound = void . Session.waitForOutbound . callChannel
+
+data ResponseKickoffException =
+    ResponseAlreadyInitiated
+  | ResponseNotInitiated
+  deriving stock (Show)
+  deriving anyclass (Exception)
+
+-- | Check if the connection to the client is healthy
+--
+-- See 'Network.GRPC.Client.isCallHealthy' for considerations regarding the
+-- non-deterministic result of this function.
+serverIsCallHealthy :: ServerCall rpc -> STM Bool
+serverIsCallHealthy = Session.isChannelHealthy . callChannel
+
+{-------------------------------------------------------------------------------
+  Internal API
+-------------------------------------------------------------------------------}
+
+-- | Send 'ProperTrailers'
+--
+-- This function is not part of the public API: we use it the top-level
+-- exception handler in "Network.GRPC.Server" to forward exceptions in server
+-- handlers to the client.
+--
+-- If no messages have been sent yet, we make use of the @Trailers-Only@ case.
+sendProperTrailers :: ServerCall rpc -> ProperTrailers -> IO ()
+sendProperTrailers ServerCall{callResponseKickoff, callChannel} trailers = do
+    updated <- atomically $ tryPutTMVar callResponseKickoff $
+                 KickoffTrailersOnly (TrailersOnly trailers)
+    unless updated $
+      -- If we didn't update, then the response has already been initiated and
+      -- we cannot make use of the Trailers-Only case.
+      atomically $ Session.send callChannel (NoMoreElems trailers)
+
 
 -- ========================================================================== --
 
@@ -411,19 +874,19 @@ clientGlobal testClock auth =
 
 serverLocal1 ::
      TestClock
-  -> Server.Call (TrivialRpc meth)
+  -> ServerCall (TrivialRpc meth)
   -> IO ()
 serverLocal1 testClock call = handle showExceptions $ do
     advanceTestClock testClock
 
     -- Wait for client early termination to become visible
     atomically $ do
-      healthy <- Server.isCallHealthy call
+      healthy <- serverIsCallHealthy call
       when healthy $ retry
 
     -- At this point, sending anything should fail
     waitForTestClockTick testClock 4
-    Server.sendOutput call $ NoMoreElems []
+    serverSendOutput call $ NoMoreElems []
   where
     showExceptions :: SomeException -> IO ()
     showExceptions err = do
@@ -432,18 +895,18 @@ serverLocal1 testClock call = handle showExceptions $ do
 
 serverLocal2 ::
      TestClock
-  -> Server.Call (TrivialRpc meth)
+  -> ServerCall (TrivialRpc meth)
   -> IO ()
 serverLocal2 testClock call = handle showExceptions $ do
     advanceTestClock testClock
 
     -- Wait for client to tell us they will send no more elements
-    _ <- Server.recvInput call
+    _ <- serverRecvInput call
     advanceTestClock testClock
 
     -- Tell the client we won't send them more elements either
     waitForTestClockTick testClock 5
-    Server.sendOutput call $ NoMoreElems []
+    serverSendOutput call $ NoMoreElems []
   where
     showExceptions :: SomeException -> IO ()
     showExceptions err = do
